로거 수정
